refactor: amend prop type file name and fix tests
